Projeto/requirements.txt
asgiref==3.4.1
cycler==0.11.0
Django==3.2.9
django-environ==0.8.1
fonttools==4.28.1
greenlet==1.1.2
gunicorn==20.1.0
kiwisolver==1.3.2
matplotlib==3.5.0
numpy==1.21.4
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
setuptools-scm==6.3.2
six==1.16.0
SQLAlchemy==1.4.27
sqlparse==0.4.2
tomli==1.2.2
whitenoise==5.3.0
